pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pptagent"
version = "0.1.0"
authors = [{ name = "Hao Zheng", email = "wszh712811@gmail.com" }]
description = "PPTAgent, a tool for utilizing LLMs to generate PowerPoint presentations from documents."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Topic :: Office/Business :: Office Suites",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = [
    "presentation",
    "powerpoint",
    "ppt",
    "llm",
    "ai",
    "document-processing",
]
dependencies = [
    "aiometer",
    "beautifulsoup4",
    "fastapi[all]",
    "einops",
    "func_argparse",
    "html2image",
    "jinja2",
    "json_repair",
    "jsonlines",
    "lxml",
    "mistune",
    "numpy",
    "oaib",
    "openai>=1.50.0",
    "opencv-python-headless",
    "pandas",
    "pdf2image",
    "peft",
    "pillow",
    "PyPDF2",
    "pydantic>=2.10.0",
    "PyYAML",
    "python-Levenshtein",
    "python-multipart",
    "python-pptx @ git+https://github.com/Force1ess/python-pptx@219513d7d81a61961fc541578c1857d08b43aa2a",
    "rich",
    "socksio",
    "tenacity",
    "tiktoken",
    "timm",
    "transformers<4.50.0",
    "uvicorn",
]

[project.urls]
"Homepage" = "https://github.com/icip-cas/PPTAgent"
"Bug Tracker" = "https://github.com/icip-cas/PPTAgent/issues"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["pptagent*"]
exclude = ["pptagent_ui"]

[tool.setuptools.package-data]
"pptagent" = ["prompts/*.txt", "roles/*.yaml", "resource/*.png"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = ["llm: mark test as using LLM (requires OPENAI_API_KEY)"]
